requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
Werkzeug==3.0.1
zipp==3.17.0
